requirements.txt
Flask==3.0.0
flask-cors==4.0.0

nltk==3.8.1

transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

python-dotenv==1.0.0
requests==2.31.0

pydantic==2.5.0
